Handle dead processes
A process could potentially have died by the time
the Chaos Monkey deems it worthy if killing (e.g.
from links to previously killed processes or simply
by completion of its task (though fear of the Chaos
Monkey seems to be the likely scenario)).

Thanks to Zabrane Mickael for the bug report.